A C implementation of Core
This should be equivalent to, but also different from, Wadler's
algorithm. Rather than representing the world as a tree of document
parts to be formatted, we just look at it as a sequence of document
parts.

To handle the "union" case we push state onto a stack, attempt to
render without newlines, and if it works clear the stack; if not we
pop the stack and retry. This is equivalent to what the functional
code is doing with "flatten" and "union" and "fits", although the
analysis can look a little more complicated there because of Haskell's
laziness.

The funny thing is that it turns out, on reflection, that we don't
really need a stack at all: the stack is only ever one deep, and
represents the point at which we started trying to flatten a group. So
this really can be crazy efficient from a memory standpoint.